fix indexing folder
