+++ src/components/Button/index.jsx
@@ -1,0 +1,1 @@
+export { default as Button } from './Button';
