+++ src/components/NavBar/index.jsx
@@ -1,0 +1,1 @@
+export { default as NavBar } from './NavBar';
