+++ src/components/OutlineTag/index.jsx
@@ -1,0 +1,1 @@
+export { default as OutlineTag } from './OutlineTag';
