+++ src/components/Profile/index.jsx
@@ -1,0 +1,1 @@
+export { default as Profile } from './Profile';
