+++ src/components/Sidebar/index.jsx
@@ -1,0 +1,1 @@
+export { default as Sidebar } from './Sidebar';
