+++ src/components/Tag/index.jsx
@@ -1,0 +1,1 @@
+export { default as Tag } from './Tag';
