+++ src/components/index.jsx
@@ -1,6 +1,6 @@
-export { default as Button } from './Button';
-export { default as Tag } from './Tag';
-export { default as OutlineTag } from './OutlineTag';
-export { default as Sidebar } from './Sidebar';
-export { default as Profile } from './Profile';
-export { default as NavBar } from './NavBar';
+export * from './Button';
+export * from './Tag';
+export * from './OutlineTag';
+export * from './Sidebar';
+export * from './Profile';
+export * from './NavBar';
